requirements.txt
Django==1.4.2
South==0.7.6
beautifulsoup4==4.1.3
requests==1.0.3
selenium==2.26.0
django-compressor==1.2
django-tastypie==0.9.11
